more instructions implemented with tests

+++ src/stack.rs
@@ -1,5 +1,5 @@
 pub struct Stack {
-    contents: Vec<u16>,
+    contents: Vec<usize>,
 }
 
 impl Stack {
@@ -7,11 +7,11 @@ impl Stack {
         Self { contents: vec![] }
     }
 
-    pub fn top(&self) -> Option<&u16> {
+    pub fn top(&self) -> Option<&usize> {
         self.contents.last()
     }
 
-    pub fn push(&mut self, val: u16) {
+    pub fn push(&mut self, val: usize) {
         self.contents.push(val);
     }
